requirements.txt
botocore==1.7.14
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
defusedxml==0.5.0
Django==1.11.5
django-allauth==0.33.0
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-logentry-admin==1.0.2
django-rest-auth==0.9.1
djangorestframework==3.6.4
docutils==0.14
greenlet==0.4.12
idna==2.6
jmespath==0.9.3
msgpack-python==0.4.8
neovim==0.1.13
oauthlib==2.0.4
pyasn1==0.3.6
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.11
six==1.11.0
sqlparse==0.2.3
unittest-xml-reporting==2.1.0
urllib3==1.22
whitenoise==3.3.1
